migrated labtest to labreport. fixed prescriptions in frontend and backend
