+++ app/src/main/webapp/WEB-INF/jsp/statistics/stats.jsp
@@ -19,7 +19,7 @@
                     <div class="panel panel-default">
                         <div class="panel-heading">
                             <h4 class="panel-title"><a data-toggle="collapse" data-parent="#accordion" href="#patients_per_specialty">
-                                Number of Patients per Doctor Speciality</a></h4>
+                                Number of Patients For Each Doctor Speciality</a></h4>
                         </div>
 
                         <div id="patients_per_specialty" class="panel-collapse collapse out">
@@ -47,7 +47,7 @@
                     <div class="panel panel-default">
                         <div class="panel-heading">
                             <h4 class="panel-title"><a data-toggle="collapse" data-parent="#accordion" href="#number_of_alerts">
-                                Number of Submitted alerts</a></h4>
+                                Number of Submitted Alerts</a></h4>
                         </div>
 
                         <div id="number_of_alerts" class="panel-collapse collapse out">
