+++ core/src/test/java/org/teamone/core/LabReportTester.java
@@ -2,23 +2,23 @@ package org.teamone.core;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.teamone.core.SQL.LabTestSQL;
+import org.teamone.core.SQL.LabReportSQL;
 import org.teamone.core.baseclasstests.TestStrings;
-import org.teamone.core.labs.LabTest;
+import org.teamone.core.labs.LabReport;
 import org.teamone.core.users.Patient;
 
 import java.util.ArrayList;
 
 import static org.junit.Assert.assertTrue;
 
-public class LabTestTester {
+public class LabReportTester {
 
-    private LabTest test;
+    private LabReport test;
 
     @Before
     public void setUp() {
 
-        test = new LabTest();
+        test = new LabReport();
         test.getPatient().setUserID(1002);
         test.setRequestionID(4);
         test.toMapObj("potassium,WHAT WHT WHAT:");
@@ -28,7 +28,7 @@ public class LabTestTester {
     @Test
     public void updateLabReport() {
         System.out.println("\nTest========update lab report");
-        test = LabTestSQL.updateLabTest(test);
+        test = LabReportSQL.updateLabReport(test);
         assertTrue("Could not update lab report", test != null);
 
         System.out.println("Update lab test successful");
@@ -42,7 +42,7 @@ public class LabTestTester {
     @Test
     public void viewLabReport() {
         System.out.println("\nTest========Viewing lab test");
-        test = LabTestSQL.viewLabTest(test);
+        test = LabReportSQL.viewLabReport(test);
         assertTrue("Could not get lab test", test != null);
         System.out.println("\nView successful");
         System.out.println("PatientID:\t" + test.getPatient().getUserID());
@@ -58,10 +58,23 @@ public class LabTestTester {
         System.out.println("\nTest========Extract all lab tests for a patient");
         Patient new1 = new Patient();
         new1.setUserID(1002);
-        ArrayList<LabTest> tempList = LabTestSQL.getListLabTestByPatient(new1);
+        ArrayList<LabReport> tempList = LabReportSQL.getListLabReportByPatient(new1);
         assertTrue("No lab tests by patient ID", !tempList.isEmpty());
-        LabTest tempRequest;
-        for (LabTest l : tempList) {
+        LabReport tempRequest;
+        for (LabReport l : tempList) {
+            System.out.println(l.getPatient().getName() + "\tTest ID: " + l.getRequestionID());
+        }
+
+        System.out.println(TestStrings.testEnd);
+
+    }
+    @Test
+    public void getAllReport() {
+
+        System.out.println("\nTest========Extract all lab tests");
+        ArrayList<LabReport> tempList = LabReportSQL.getAllLabReports();
+        assertTrue("No lab tests ", !tempList.isEmpty());
+        for (LabReport l : tempList) {
             System.out.println(l.getPatient().getName() + "\tTest ID: " + l.getRequestionID());
         }
